add compile argument to load_model
add compilation argument to load_model in /models/__init__.py to remove the message: "WARNING:tensorflow:No training configuration found in the save file, so the model was *not* compiled. Compile it manually.". This message appears because tensorflow can't store the optimizer algorithm and when loading them it makes a reminder that it's not added, but we only use the model for inference so it's not needed.